Version 1
When in a group and playing YouTube videos it always happens that the
person controlling the device connected to the speakers gets to play all
the songs he wants. But when others want their favorite songs to be
played they have to request that person to play their song.

This app is developed with this problem in mind. It solves the problem
by giving every person in the group the chance to upload the song to a
central playlist. The central playlist is created with each user's
favorite song added to the list. The device connected to the speakers
will pick songs from the playlist and play them in the sequence.

Every user right now gets to upload 1 song per playlist. After the
entire playlist has played the uploading can begin again with each group
member can begin to add their favorite YouTube video to the playlist.